requirements.txt
# Core Web Framework
fastapi==0.95.1
uvicorn==0.22.0

# WhatsApp and Messaging
twilio==8.2.0
requests==2.30.0  # For Instagram/Meta API calls

# AI and NLP
openai==0.27.6
langdetect==1.0.9

# Payment Processing
stripe==5.4.0

# Database
sqlite3

# Dashboard
streamlit==1.22.0

# Utilities
python-dotenv==1.0.0
pydantic==1.10.7

# Optional: Logging and Monitoring
loguru==0.7.0 